Update button build fix
