+++ src/components/Misc/Settings/SettingsBuildingCodes.vue
@@ -64,38 +64,14 @@ import type { dropdownItem } from '@/components/Misc/DropdownMenuItem.vue'
 import { buildingCodes } from '@/models/buildingCode'
 
 import Dropdown from '@/components/Misc/Dropdown.vue'
+import UpdateButton from '@/components/Misc/Settings/UpdateButton.vue'
 import { storeToRefs } from 'pinia'
 
 export default defineComponent({
   name: 'SettingsBuildingCodes',
   components: {
     Dropdown,
-    // Local update button
-    UpdateButton: defineComponent({
-      name: 'UpdateButton',
-      props: {
-        label: {
-          type: String,
-          default: 'Update',
-        },
-      },
-      setup(props, { emit }) {
-        const handleClick = () => {
-          emit('click')
-        }
-
-        return { handleClick }
-      },
-      template: `
-        <button 
-          type="button" 
-          class="font-semibold text-green-600 hover:text-green-500"
-          @click="handleClick"
-        >
-          {{ label }}
-        </button>
-      `,
-    }),
+    UpdateButton
   },
   setup() {
     const settingsStore = useSettingsStore()
