+++ src/components/Misc/Settings/SettingsLifecycleStages.vue
@@ -40,39 +40,15 @@
 <script lang="ts">
 import { defineComponent, ref, reactive, watch } from 'vue'
 import { useSettingsStore } from '@/stores/settings'
+import UpdateButton from '@/components/Misc/Settings/UpdateButton.vue'
 
 import { LifeCycleStages } from '@/models/material'
-import { storeToRefs } from 'pinia';
+import { storeToRefs } from 'pinia'
 
 export default defineComponent({
   name: 'SettingsLifecycleStages',
   components: {
-    // Local update button
-    UpdateButton: defineComponent({
-      name: 'UpdateButton',
-      props: {
-        label: {
-          type: String,
-          default: 'Update',
-        },
-      },
-      setup(props, { emit }) {
-        const handleClick = () => {
-          emit('click')
-        }
-
-        return { handleClick }
-      },
-      template: `
-        <button 
-          type="button" 
-          class="font-semibold text-green-600 hover:text-green-500"
-          @click="handleClick"
-        >
-          {{ label }}
-        </button>
-      `,
-    }),
+    UpdateButton
   },
   setup() {
     const settingsStore = useSettingsStore()
